pyproject.toml
[project]
name = "mqtt-alertd"
version = "0.1.0"
description = "MQTT Daemon to act on ops alerts"
readme = "README.md"
authors = [
    { name = "Laura Batalha", email = "5883822+lbatalha@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "docker>=7.1.0",
    "paho-mqtt>=2.1.0",
    "pyyaml>=6.0.2",
]

[project.scripts]
mqtt-alertd = "mqtt_alertd:main"

[build-system]
requires = ["uv_build>=0.8.0,<0.9"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "setuptools>=80.9.0",
]
